// src/modules/route-engine/enums/route.enum.ts
// src/modules/route-engine/enums/route.enum.ts

export enum RouteType {
  IDEA_NEGOCIO = 'IDEA_NEGOCIO',
  NEGOCIO_EXISTENTE = 'NEGOCIO_EXISTENTE',
  INVERSION = 'INVERSION',
  ACTIVO_DIGITAL = 'ACTIVO_DIGITAL',
  FRANQUICIA = 'FRANQUICIA',
  EXPANSION = 'EXPANSION',
  COMPRA_EMPRESA = 'COMPRA_EMPRESA',
  COMPARACION = 'COMPARACION',
}